split queries into domain modules
